removed unused server routes
