+++ server/routers/eventsRouter.ts
@@ -1,78 +1,44 @@
-import Router from 'express';
-import redisMonitors from '../redis-monitors/redis-monitors';
+import express from 'express';
+import monitorsController from '../controllers/monitorsController';
+import eventsController from '../controllers/eventsController';
 
-const router = Router();
+const router = express.Router();
 
-router.get('/', (req, res, next) => {
-  //set const body = {data: []}
-  const body = { data: [] };
-  //if not, then loop through eventsMonitors 
-  redisMonitors.forEach(monitor => {
-    //for each monitor, create a new object "instance" (to be put in the response.body)
-    const instance = {
-      //instance.instanceId = monitor.instanceId 
-      instanceId: monitor.instanceId,
-      //instance.keyspaces = [];
-      keyspaces: []
-    };
-    //loop through monitor.keyspaces array
-    monitor.keyspaces.forEach(keyspace => {
-      //for each keyspace, grab keyspace.eventLog, convert into array using returnLogAsArray and push into instance.keyspaces
-      //check if req.query.eventTotal != undefined, if so use as argument for returnLogAsArray
-      const { eventTotal } = req.query;
-      instance.keyspaces.push(keyspace.eventLog.returnLogAsArray((+eventTotal) ? +eventTotal : 0))
-    })
-    //push instance object into body.data
-    body.data.push(instance)
-  })
-  //res.send jsonified body.data
-  res.status(200).json(body);
+router.get('/',
+  monitorsController.findAllMonitors,
+  eventsController.refreshEventLog,
+  eventsController.getEventsPages,
+  (req: express.Request, res:express.Response): void => {
+    res.status(200).json(res.locals.events);
 });
 
-router.get('/:instanceId/', (req, res, next) => {
-  //set const body = {data: []}
-  const body = { data: [] };
-  const { instanceId } = req.params
-  //find monitor based on instanceId
-  const monitor = redisMonitors.find(m => {
-    return m.instanceId === +instanceId;
-  });
-  //initialize object instance as in previous section
-  const instance = {
-    instanceId: monitor.instanceId,
-    keyspaces: []
-  };
-  //loop through monitor.keyspaces array
-  monitor.keyspaces.forEach(keyspace => {
-    //grab keyspace.eventLog, convert into array and push into instance.keyspaces
-    const { eventTotal } = req.query;
-    instance.keyspaces.push(keyspace.eventLog.returnLogAsArray((+eventTotal) ? +eventTotal : 0))
-  });
-
-  body.data.push(instance);
-  //res.send jsonified body.data
-  res.status(200).json(body);
+router.get('/:instanceId',
+  monitorsController.findSingleMonitor,
+  eventsController.refreshEventLog,
+  eventsController.getEventsPages,
+  (req: express.Request, res:express.Response): void => {
+    res.status(200).json(res.locals.events);
 });
 
-router.get('/:instanceId/:dbIndex', (req, res, next) => {
-  //set const body = {data: []}
-  const body = { data: [] };
-  const { instanceId, dbIndex } = req.params
-  //find monitor based on instanceId
-  const monitor = redisMonitors.find(m => {
-    return m.instanceId === +instanceId;
-  });
-  //initialize object instance as in previous section
-  const instance = {
-    instanceId: monitor.instanceId,
-    keyspaces: []
-  };
-  //grab keyspace.eventLog, convert into array and push into instance.keyspaces
-  const { eventTotal } = req.query;
-  instance.keyspaces.push(monitor.keyspaces[dbIndex].eventLog.returnLogAsArray((eventTotal) ? eventTotal : 0))
-  body.data.push(instance);
-  //res.send jsonified body.data
-  res.status(200).json(body);
+
+router.get('/:instanceId/:dbIndex',
+  monitorsController.findSingleMonitor,
+  eventsController.refreshEventLog,
+  eventsController.getEventsPages,
+  (req: express.Request, res:express.Response): void => {
+    res.status(200).json(res.locals.events);
+});
+
+router.get('/totals/:instanceId/:dbIndex',
+  monitorsController.findSingleMonitor,
+  eventsController.validateRequestType,
+  (req: express.Request, res: express.Response, next: express.NextFunction): void => {
+    if (req.query.timeInterval) eventsController.getEventsByTimeInterval(req, res, next);
+    else if (req.query.eventTotal) eventsController.getSingleEventsTotal(req, res, next);
+  },  
+  (req: express.Request, res: express.Response): void => {
+    res.status(200).json(res.locals.eventTotals);
 });
 
 export default router;
+
